RegressionFormulae.jl

Base.:/

Base.:^

Base.:/ — Method

group / term

Generate predictors for term within each level of group. Implemented as group + fulldummy(group) & term.

Base.:^ — Method

(term1, term2, ...) ^ n

Generate all interactions of terms up to order $n$.

Warning

Embedded InteractionTerms (i.e. (a + b + c & d)^2) are not currently supported and result in an error.